requirements.txt
requests==2.31.0
Jinja2==3.1.2
python-dotenv==1.0.0
Pillow==10.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
tomli==2.0.1
# Gateway Dependencies
fastapi
uvicorn
pydantic
websockets
markdown
mdit-py-plugins
python-multipart
